requirements.txt
numpy==1.26.4
pickle-mixin==1.0.2
streamlit==1.31.0
seaborn==0.13.2
pandas==2.2.2
matplotlib==3.10.0
scikit-learn==1.6.1
googletrans==4.0.0-rc1
google-generativeai==0.8.4
